Moved GestureHandleRootView to App.js
